fix: move CSVFile into SettingsView.swift to resolve scope error
CSVFile was defined in LocationExporter.swift which may not be registered
in the Xcode project if xcodeproj was generated before the file was added.
Moving it to SettingsView.swift (its only use site) fixes the build.
