+++ PlaceNotes/Services/LocationExporter.swift
@@ -1,6 +1,4 @@
 import Foundation
-import SwiftUI
-import UniformTypeIdentifiers
 
 enum LocationExporter {
     private static let iso8601: ISO8601DateFormatter = {
@@ -30,23 +28,5 @@ enum LocationExporter {
         }
 
         return lines.joined(separator: "\n").data(using: .utf8) ?? Data()
-    }
-}
-
-struct CSVFile: FileDocument {
-    static var readableContentTypes: [UTType] { [.commaSeparatedText] }
-
-    var data: Data
-
-    init(data: Data) {
-        self.data = data
-    }
-
-    init(configuration: ReadConfiguration) throws {
-        self.data = configuration.file.regularFileContents ?? Data()
-    }
-
-    func fileWrapper(configuration: WriteConfiguration) throws -> FileWrapper {
-        FileWrapper(regularFileWithContents: data)
     }
 }
